+++ domain/dropbox/service/sv_file_url/url_test.go
@@ -3,38 +3,14 @@ package sv_file_url
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_path"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
-	"go.uber.org/zap"
-	"strings"
 	"testing"
 )
 
 const (
 	DummyImageUrl = "https://dummyimage.com/64x64/888/222.png"
 )
-
-func TestUrlImpl_SaveWithTestSuite(t *testing.T) {
-	qt_api.DoTestTokenFull(func(ctx dbx_context.Context) {
-		svc := New(ctx)
-		path := qt_api.ToolboxTestSuiteFolder.ChildPath("save_url", "f0.png")
-		entry, err := svc.Save(path, DummyImageUrl)
-		if err != nil {
-			t.Error(err.Error())
-			return
-		}
-		if entry.Tag() != "file" {
-			t.Error("invalid")
-		}
-
-		// file name might be auto renamed into like `f0 (1).png`, if duplicated file there
-		if !strings.HasPrefix(entry.Name(), "f0") {
-			t.Error("invalid")
-		}
-		ctx.Log().Debug("entry", zap.Any("entry", entry))
-	})
-}
 
 func TestPathWithName(t *testing.T) {
 	base := mo_path.NewDropboxPath("/Test/Path")