+++ domain/dropbox/service/sv_namespace/namespace_test.go
@@ -2,33 +2,10 @@ package sv_namespace
 
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
-	"go.uber.org/zap"
 	"testing"
 )
-
-func TestEndToEndNamespaceImpl_List(t *testing.T) {
-	qt_api.DoTestBusinessFile(func(ctx dbx_context.Context) {
-		svc := newTest(ctx, 3)
-		namespaces, err := svc.List()
-		if err != nil {
-			t.Error(err)
-			return
-		}
-		if len(namespaces) < 1 {
-			t.Error("invalid")
-			return
-		}
-		for _, n := range namespaces {
-			if n.NamespaceId == "" {
-				t.Error("invalid")
-			}
-			ctx.Log().Debug("namespace", zap.Any("namespace", n))
-		}
-	})
-}
 
 // Mock tests