+++ domain/dropbox/service/sv_sharedfolder_mount/mount_test.go
@@ -3,28 +3,10 @@ package sv_sharedfolder_mount
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_sharedfolder"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
 	"testing"
 )
-
-func TestEndToEndMountImpl_List(t *testing.T) {
-	qt_api.DoTestTokenFull(func(ctx dbx_context.Context) {
-		svc := New(ctx)
-		mounts, err := svc.List()
-		if err != nil {
-			t.Error(err)
-			return
-		}
-
-		for _, m := range mounts {
-			if m.SharedFolderId == "" || m.Name == "" {
-				t.Error("invalid")
-			}
-		}
-	})
-}
 
 // mock test