+++ domain/dropbox/service/sv_sharedlink/sharedlink_test.go
@@ -4,104 +4,11 @@ import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_sharedlink"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_url"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
-	"strings"
 	"testing"
 	"time"
 )
-
-func TestEndToEndSharedLinkImpl_CreateListRemove(t *testing.T) {
-	qt_api.DoTestTokenFull(func(ctx dbx_context.Context) {
-		svc := New(ctx)
-		src := qt_api.ToolboxTestSuiteFolder.ChildPath("copy/F0.jpg")
-		links, err := svc.ListByPath(src)
-		if err != nil {
-			t.Error(err)
-			return
-		}
-
-		// clean up existing link
-		for _, link := range links {
-			if err := svc.Remove(link); err != nil {
-				t.Error(err)
-			}
-		}
-
-		// Default link
-		{
-			link, err := svc.Create(src)
-			if err != nil {
-				t.Error(err)
-				return
-			}
-			if link.LinkVisibility() != "public" {
-				t.Error("invalid visibility")
-			}
-			if strings.ToLower(link.LinkName()) != "f0.jpg" {
-				t.Error("invalid name")
-			}
-			if link.LinkExpires() != "" {
-				t.Error("invalid expire")
-			}
-
-			// clean up
-			err = svc.Remove(link)
-			if err != nil {
-				t.Error(err)
-			}
-		}
-
-		// Team only link
-		{
-			link, err := svc.Create(src, TeamOnly())
-			if err != nil {
-				t.Error(err)
-				return
-			}
-			if link.LinkVisibility() != "team_only" {
-				t.Error("invalid visibility")
-			}
-			if strings.ToLower(link.LinkName()) != "f0.jpg" {
-				t.Error("invalid name")
-			}
-			if link.LinkExpires() != "" {
-				t.Error("invalid expire")
-			}
-
-			// clean up
-			err = svc.Remove(link)
-			if err != nil {
-				t.Error(err)
-			}
-		}
-
-		// Password protected link
-		{
-			link, err := svc.Create(src, Password("toolbox!"))
-			if err != nil {
-				t.Error(err)
-				return
-			}
-			if link.LinkVisibility() != "password" {
-				t.Error("invalid visibility")
-			}
-			if strings.ToLower(link.LinkName()) != "f0.jpg" {
-				t.Error("invalid name")
-			}
-			if link.LinkExpires() != "" {
-				t.Error("invalid expire")
-			}
-
-			// clean up
-			err = svc.Remove(link)
-			if err != nil {
-				t.Error(err)
-			}
-		}
-	})
-}
 
 // Mock tests