+++ domain/dropbox/service/sv_teamfolder/teamfolder_test.go
@@ -3,35 +3,10 @@ package sv_teamfolder
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
 	"github.com/watermint/toolbox/domain/dropbox/model/mo_teamfolder"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
 	"testing"
 )
-
-func TestEndToEndTeamFolderImpl_List(t *testing.T) {
-	qt_api.DoTestBusinessFile(func(ctx dbx_context.Context) {
-		svc := New(ctx)
-		list, err := svc.List()
-		if err != nil {
-			t.Error(err)
-			return
-		}
-
-		for _, tf := range list {
-			if tf.TeamFolderId == "" {
-				t.Error("invalid")
-			}
-			r, err := svc.Resolve(tf.TeamFolderId)
-			if err != nil {
-				t.Error(err)
-			}
-			if r.TeamFolderId != tf.TeamFolderId || r.Name != tf.Name {
-				t.Error("invalid")
-			}
-		}
-	})
-}
 
 // Mock test