+++ domain/dropbox/service/sv_usage/usage_test.go
@@ -2,20 +2,10 @@ package sv_usage
 
 import (
 	"github.com/watermint/toolbox/domain/dropbox/api/dbx_context"
-	"github.com/watermint/toolbox/quality/infra/qt_api"
 	"github.com/watermint/toolbox/quality/infra/qt_errors"
 	"github.com/watermint/toolbox/quality/infra/qt_recipe"
 	"testing"
 )
-
-func TestEndToEndUsageImpl_Resolve(t *testing.T) {
-	qt_api.DoTestTokenFull(func(ctx dbx_context.Context) {
-		_, err := New(ctx).Resolve()
-		if err != nil {
-			t.Error(err)
-		}
-	})
-}
 
 // Mock tests